refactor: useComments 를 comment entity로 이동
